youtube-outline/ingest-transcripts/requirements.txt
boto3>=1.26.0
deepgram-sdk>=2.11.0
python-dotenv>=0.19.0
pytube>=15.0.0
requests>=2.31.0
